+++ src/components/Project/ProjectToolsDrawer/TransferOwnershipForm.tsx
@@ -1,19 +1,14 @@
 import { t, Trans } from '@lingui/macro'
 import { Button, Form } from 'antd'
 import { EthAddressInput } from 'components/inputs/EthAddressInput'
-import { TransactorInstance } from 'hooks/Transactor'
-import { useState } from 'react'
+import { V2ProjectContext } from 'contexts/v2/projectContext'
+import { useTransferProjectOwnershipTx } from 'hooks/v2/transactor/TransferProjectOwnershipTx'
+import { useContext, useState } from 'react'
 import { emitErrorNotification } from 'utils/notifications'
 
-export function TransferOwnershipForm({
-  useTransferProjectOwnershipTx,
-  ownerAddress,
-}: {
-  useTransferProjectOwnershipTx: () => TransactorInstance<{
-    newOwnerAddress: string
-  }>
-  ownerAddress: string | undefined
-}) {
+export function TransferOwnershipForm() {
+  const { projectOwnerAddress } = useContext(V2ProjectContext)
+
   const transferProjectOwnershipTx = useTransferProjectOwnershipTx()
 
   const [loadingTransferOwnership, setLoadingTransferOwnership] =
@@ -47,7 +42,7 @@ export function TransferOwnershipForm({
         <Trans>Transfer ownership</Trans>
       </h3>
       <p>
-        <Trans>Current owner: {ownerAddress}</Trans>
+        <Trans>Current owner: {projectOwnerAddress}</Trans>
       </p>
       <Form.Item name="to" label={t`Recipient address`}>
         <EthAddressInput />
