--- src/components/v2/V2Project/V2ProjectSettingsPage/V2ProjectSettingsPayoutsContent.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-const V2ProjectSettingsPayoutsContent = () => {
-  return <div>V2ProjectSettingsPayoutsContent</div>
-}
-
-export default V2ProjectSettingsPayoutsContent
