+++ src/components/veNft/VeNftEnableSection.tsx
@@ -2,25 +2,24 @@ import { Trans } from '@lingui/macro'
 import { Button, Divider, Space } from 'antd'
 import { useContext, useState } from 'react'
 
-import VeNftSetupModal from 'components/veNft/VeNftSetupModal'
 import { useLaunchVeNftTx } from 'hooks/veNft/transactor/LaunchVeNftTx'
 import { NetworkContext } from 'contexts/networkContext'
 import { V2ProjectContext } from 'contexts/v2/projectContext'
 
-import { featureFlagEnabled } from 'utils/featureFlags'
+import FormattedAddress from 'components/FormattedAddress'
 
 import {
   DEFAULT_LOCK_DURATIONS,
   VENFT_RESOLVER_ADDRESS,
 } from 'constants/veNft/veNftProject'
-import { FEATURE_FLAGS } from 'constants/featureFlags'
 
 const VeNftEnableSection = () => {
   const { userAddress, onSelectWallet } = useContext(NetworkContext)
-  const { tokenSymbol } = useContext(V2ProjectContext)
+  const {
+    tokenSymbol,
+    veNft: { contractAddress: veNftContractAddress },
+  } = useContext(V2ProjectContext)
   const [loading, setLoading] = useState(false)
-  const [setupModalVisible, setSetupModalVisible] = useState(false)
-  const veNftCreatorEnabled = featureFlagEnabled(FEATURE_FLAGS.VENFT_CREATOR)
 
   const launchVeBannyTx = useLaunchVeNftTx()
 
@@ -63,43 +62,24 @@ const VeNftEnableSection = () => {
               veBanny assets and parameters.
             </Trans>
           </p>
-          <Button
-            type="primary"
-            size="small"
-            onClick={launchVeBanny}
-            loading={loading}
-          >
-            Launch veBanny
-          </Button>
-        </section>
-        <Divider />
-        {veNftCreatorEnabled && (
-          <section>
-            <h3>
-              <Trans>Enable veNFT Governance</Trans>
-            </h3>
-            <p>
-              <Trans>
-                Set up and launch veNFT governance for your project using custom
-                assets and parameters. (Experimental, does not currently launch)
-              </Trans>
-            </p>
+          {!veNftContractAddress ? (
             <Button
               type="primary"
               size="small"
-              onClick={() => setSetupModalVisible(true)}
+              onClick={launchVeBanny}
+              loading={loading}
             >
-              Launch Setup
+              <Trans>Launch veBanny</Trans>
             </Button>
-          </section>
-        )}
+          ) : (
+            <span>
+              <Trans>Contract address:</Trans>{' '}
+              <FormattedAddress address={veNftContractAddress} />
+            </span>
+          )}
+        </section>
+        <Divider />
       </>
-      {setupModalVisible && (
-        <VeNftSetupModal
-          visible={setupModalVisible}
-          onCancel={() => setSetupModalVisible(false)}
-        />
-      )}
     </Space>
   )
 }
