+++ src/locales/messages.pot
@@ -449,6 +449,9 @@ msgstr ""
 msgid "Available:"
 msgstr ""
 
+msgid "Back to project"
+msgstr ""
+
 msgid "Back to top"
 msgstr ""
 
@@ -650,6 +653,9 @@ msgstr ""
 msgid "Connected wallet not authorized"
 msgstr ""
 
+msgid "Contract address:"
+msgstr ""
+
 msgid "Contract address: {0}"
 msgstr ""
 
@@ -725,7 +731,7 @@ msgstr ""
 msgid "Current 3rd Party Exchange Rates"
 msgstr ""
 
-msgid "Current owner: {ownerAddress}"
+msgid "Current owner: {projectOwnerAddress}"
 msgstr ""
 
 msgid "Currently worth: <0><1/></0>"
@@ -1002,9 +1008,6 @@ msgid "Edit tracked assets"
 msgstr ""
 
 msgid "Enable veBanny Governance"
-msgstr ""
-
-msgid "Enable veNFT Governance"
 msgstr ""
 
 msgid "Enable veNFT to Spend Unclaimed Tokens"
@@ -1329,6 +1332,9 @@ msgid "Latest payments"
 msgstr ""
 
 msgid "Launch funding cycle"
+msgstr ""
+
+msgid "Launch veBanny"
 msgstr ""
 
 msgid "Launch veNFT"
@@ -2115,9 +2121,6 @@ msgid "Set this to the sum of all your payouts"
 msgstr ""
 
 msgid "Set up V1 token migration"
-msgstr ""
-
-msgid "Set up and launch veNFT governance for your project using custom assets and parameters. (Experimental, does not currently launch)"
 msgstr ""
 
 msgid "Set up and launch veNFT governance for your project using default veBanny assets and parameters."
